+++ .gitignore
@@ -74,11 +74,10 @@ web_modules/
 
 # dotenv environment variable files
 .env
-.env.local
 .env.development.local
 .env.test.local
 .env.production.local
-.env*.local
+.env.local
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
@@ -129,13 +128,3 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
-
-# IDEs
-.idea
-.vscode
-*.swp
-*.swo
-
-# OS files
-.DS_Store
-Thumbs.db
