+++ README.md
@@ -3,12 +3,10 @@
 <p> MU Calendar Sync is a web app to sync your MUERP timetable to your Google Calendar, iCal, or any other calendar format. </p>
 
 <p>
-    <img src="https://img.shields.io/badge/next.js-000000?style=for-the-badge&logo=nextdotjs" alt="Next.js">
+    <img alt="Static Badge" src="https://img.shields.io/badge/next.js-000000?style=for-the-badge&logo=nextdotjs" alt="Next.js">
     <img src="https://img.shields.io/badge/typescript-%23007ACC.svg?style=for-the-badge&logo=typescript&logoColor=white" alt="TypeScript">
-    <img src="https://img.shields.io/badge/python-3670A0?style=for-the-badge&logo=python&logoColor=ffdd54" alt="Python">
+    <img alt="Static Badge" src="https://img.shields.io/badge/python-3670A0?style=for-the-badge&logo=python&logoColor=ffdd54" alt="Python">
     <img src="https://img.shields.io/badge/vercel-%23000000.svg?style=for-the-badge&logo=vercel&logoColor=white" alt="Vercel">
-    <img src="https://img.shields.io/badge/Heroku-430098?style=for-the-badge&logo=heroku&logoColor=fffe" alt="Heroku">
-    <img src="https://img.shields.io/badge/MongoDB-%234ea94b.svg?style=for-the-badge&logo=mongodb&logoColor=white" alt="MongoDB">
 </p>
 </div>
 
@@ -21,7 +19,7 @@
 ## Usage and Features
 
 Simply enter your MUERP credentials and the app will sync your timetable to your
-selected calendar. This repository only has the front end of the web app. For backend, [contact me](mailto:mark.atharv@gmail.com).
+selected calendar.
 
 ## How to install
