+++ package.json
@@ -13,8 +13,6 @@
   "dependencies": {
     "@auth/core": "^0.34.2",
     "@tabler/icons-react": "^3.21.0",
-    "@types/crypto-js": "^4.2.2",
-    "crypto-js": "^4.2.0",
     "next": "15.0.2",
     "next-auth": "^4.24.10",
     "next-themes": "^0.3.0",
@@ -36,7 +34,7 @@
     "tailwindcss": "^3.4.14",
     "typescript": "^5.6.3"
   },
-  "description": "MU Calendar Sync is a web app to sync your MUERP timetable to your Google Calendar, iCal, or any other calendar format.",
+  "description": "This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-app`](https://nextjs.org/docs/app/api-reference/cli/create-next-app).",
   "main": "index.js",
   "author": "Atharv Garg <markvi2>",
   "license": "MIT",
