+++ src/utils/authOptions.ts
@@ -1,41 +1,1 @@
-import GoogleProvider from "next-auth/providers/google";
-import type { NextAuthOptions } from "next-auth";
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-      authorization: {
-        params: {
-          prompt: "consent",
-          access_type: "offline",
-          scope:
-            "https://www.googleapis.com/auth/calendar.events https://www.googleapis.com/auth/calendar openid email profile",
-        },
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, account, user }) {
-      if (account && user) {
-        return {
-          ...token,
-          accessToken: account.access_token,
-          refreshToken: account.refresh_token,
-          accessTokenExpires: account.expires_at
-            ? account.expires_at * 1000
-            : 0,
-        };
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.accessToken = token.accessToken as string;
-      }
-      return session;
-    },
-  },
-  debug: process.env.NODE_ENV === "development",
-};
+
