--- src/utils/encryption.ts
@@ -1,10 +1,0 @@
-import CryptoJS from "crypto-js";
-
-export const encryptData = (data: any, key: string) => {
-  return CryptoJS.AES.encrypt(JSON.stringify(data), key).toString();
-};
-
-export const decryptData = (encryptedData: string, key: string) => {
-  const bytes = CryptoJS.AES.decrypt(encryptedData, key);
-  return JSON.parse(bytes.toString(CryptoJS.enc.Utf8));
-};
